+++ src/utils/start-button.js
@@ -1,0 +1,7 @@
+AFRAME.registerComponent("start-button", {
+  init: function () {
+    this.el.addEventListener("click", () => {
+      AFRAME.scenes[0].emit("showBaseline");
+    });
+  },
+});
